refactor: permissions package structure in androidMain
