+++ identity_presentation/src/androidMain/kotlin/net/thechance/mena/identity/presentation/util/permissions/CameraPermission.kt
@@ -1,13 +1,13 @@
-package net.thechance.mena.identity.presentation.util
+package net.thechance.mena.identity.presentation.util.permissions
 
 import android.Manifest
 import android.content.Context
 import android.content.pm.PackageManager
-import net.thechance.mena.identity.domain.exception.PermissionDeniedException
-import net.thechance.mena.identity.domain.exception.PermissionDeniedPermanentlyException
-import net.thechance.mena.identity.domain.exception.PermissionNotDeterminedException
+import net.thechance.mena.identity.presentation.util.PermissionManager
+import net.thechance.mena.identity.presentation.util.permissions.util.openAppSettingsPage
 import net.thechance.mena.identity.presentation.util.permissionHandler.PermissionController
 import net.thechance.mena.identity.presentation.util.permissionHandler.PermissionState
+import net.thechance.mena.identity.presentation.util.permissions.util.handlePermissionState
 
 internal class CameraPermission(
     private val permissionManager: PermissionManager,
@@ -30,15 +30,6 @@ internal class CameraPermission(
 
     override suspend fun requestPermission() {
         permissionManager.requestPermission(listOf(requiredPermission))
-            .values.forEach(::handlePermissionResult)
-    }
-
-    private fun handlePermissionResult(state: PermissionState) {
-        when (state) {
-            PermissionState.NOT_DETERMINED -> throw PermissionNotDeterminedException()
-            PermissionState.GRANTED -> {}
-            PermissionState.DENIED -> throw PermissionDeniedException()
-            PermissionState.DENIED_PERMANENTLY -> throw PermissionDeniedPermanentlyException()
-        }
+            .values.forEach(::handlePermissionState)
     }
 }
